+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Added saturation to joint velocity, joint position, and joint impedance examples
   * Changed examples to read initial states inside control loops (after controller switching)
   * Always throw `ControlException`s for control-related command responses
+  * Add support for commanding elbow positions for Cartesian motions
 
 ## 0.1.0 - 2017-09-15